requirements.txt
allure-pytest==2.13.5
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
et_xmlfile==2.0.0
Flask==3.1.2
greenlet==3.3.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
MouseInfo==0.1.3
numpy==2.2.6
opencv-python==4.12.0.88
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.1.0
playwright==1.57.0
pymysql==1.1.0
cryptography==46.0.3
pytest==9.0.2
PyAutoGUI==0.9.54
pycparser==2.23
pyee==13.0.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pypdf==6.5.0
PyPDF2==3.0.1
pypdfium2==5.2.0
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest-base-url==2.1.0
pytest-playwright==0.7.2
python-docx==1.2.0
python-dotenv==1.2.1
python-slugify==8.0.4
pytweening==1.2.0
requests==2.31.0
selenium==4.39.0
Send2Trash==2.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
text-unidecode==1.3
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.6.2
webdriver-manager==4.0.2
websocket-client==1.9.0
Werkzeug==3.1.4
wsproto==1.3.2
celery==5.3.6
redis==5.0.1
responses==0.23.3
prometheus-flask-exporter==0.23.0
Flask-Login==0.6.3
locust==2.33.0
